Switching to a much simpler design.
There is no point in over-templating the ipaca interface. It is much easier
to define our own datatypes in the detail namespace and then only provide
templates for the conversion operators.
The code now provides basic datatypes for a mass spectrum and a stoichiometry.

The mercury code was refactored into Mercury7; isotope pattern calculation
is now not responsible for charge adjustments anymore. This should be
provided through suitable warpper functions.

The code still requires some testing.

Signed-off-by: Marc Kirchner <mail@marc-kirchner.de>