Moved the Mercury7 implementation into an implementation class.
We now provide an interface for arbitrary, user-defined stoichiometry
and spectrum types. All the user needs to to is to provide a conversion
function. This is acceptable because it guarantees near-C speed inside
the lib (where all the convolutions are done) and requires only a
single copy in and a single copy out.

Signed-off-by: Marc Kirchner <mail@marc-kirchner.de>