Enabled calculation in multiple charges.
The implementation follows a simple principle: we adjust the number of
hydrogen atoms (if necessary), calculate the uncharged isotope
distribution and then adjust m/z assuming that the remaining mass
corrections come from electrons. This yields a simple logic and
keeps the charge logic out of the isotope pattern calculation.

The interface is still a bit hacky; we need to see a few application
scenarios before deciding how to stabilize it.

Signed-off-by: Marc Kirchner <mail@marc-kirchner.de>